Rename "format.py" to "formats.py"
To remain consistent with this naming everywhere.
