+++ src/elapi/styles/__init__.py
@@ -2,6 +2,6 @@
 from ._markdown_doc import get_custom_help_text
 from ._missing import Missing
 from .base import stdin_console, stderr_console
-from .format import BaseFormat, FormatError, Format, ValidateLanguage
+from .formats import BaseFormat, FormatError, Format, ValidateLanguage
 from .highlight import BaseHighlight, Highlight, NoteText, ColorText, print_typer_error
 from .rich_utils import rich_format_help_with_callback
